metadata: Move validation hashing to dedicated file
